refactor: code improve
